add ModeratorActions to SunshineNewUsersInfo
