+++ packages/lesswrong/components/sunshineDashboard/ModeratorActions.tsx
@@ -1,22 +1,63 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { registerComponent, Components } from '../../lib/vulcan-lib';
 import DoneIcon from '@material-ui/icons/Done';
 import SnoozeIcon from '@material-ui/icons/Snooze';
 import AddAlarmIcon from '@material-ui/icons/AddAlarm';
 import DeleteForeverIcon from '@material-ui/icons/DeleteForever';
 import RemoveCircleOutlineIcon from '@material-ui/icons/RemoveCircleOutline';
+import VisibilityOutlinedIcon from '@material-ui/icons/VisibilityOutlined';
 import OutlinedFlagIcon from '@material-ui/icons/OutlinedFlag';
 import classNames from 'classnames';
 import { useUpdate } from '../../lib/crud/withUpdate';
 import { useCreate } from '../../lib/crud/withCreate';
 import moment from 'moment';
-import { RATE_LIMIT_ONE_PER_DAY } from '../../lib/collections/moderatorActions/schema';
+import { MODERATOR_ACTION_TYPES, RATE_LIMIT_ONE_PER_DAY } from '../../lib/collections/moderatorActions/schema';
 import FlagIcon from '@material-ui/icons/Flag';
+import Input from '@material-ui/core/Input';
 
 const styles = (theme: ThemeType): JssStyles => ({
-  root: {
-
-  }
+  row: {
+    display: "flex",
+    alignItems: "center",
+    marginTop: 8,
+    marginBottom: 8
+  },
+  modButton:{
+    marginTop: 6,
+    marginRight: 16,
+    cursor: "pointer",
+    '&:hover': {
+      opacity: .5
+    }
+  },
+  snooze10: {
+    color: theme.palette.primary.main,
+    fontSize: 34,
+    marginTop: 4
+  },
+  permissionsButton: {
+    padding: 6,
+    paddingTop: 3,
+    paddingBottom: 3,
+    border: theme.palette.border.normal,
+    borderRadius: 2,
+    cursor: "pointer",
+    whiteSpace: "nowrap",
+    marginRight: 8
+  },
+  permissionDisabled: {
+    border: "none"
+  },
+  notes: {
+    border: theme.palette.border.faint,
+    borderRadius: 2,
+    paddingLeft: 8,
+    paddingRight: 8,
+    paddingTop: 4,
+    paddingBottom: 4,
+    marginTop: 8,
+    marginBottom: 8,
+  },
 });
 
 interface UserContentCountPartial {
@@ -34,13 +75,14 @@ export function getNewSnoozeUntilContentCount(user: UserContentCountPartial, con
   return getCurrentContentCount(user) + contentCount
 }
 
-export const ModeratorActions = ({classes, user, currentUser, refetch, setNotes, }: {
+export const ModeratorActions = ({classes, user, currentUser, refetch}: {
   user: SunshineUsersList,
   classes: ClassesType,
   currentUser: UsersCurrent,
-  refetch: () => void
+  refetch?: () => void
 }) => {
   const { LWTooltip } = Components
+  const [notes, setNotes] = useState(user.sunshineNotes || "")
 
   const { mutate: updateUser } = useUpdate({
     collectionName: "Users",
@@ -68,7 +110,31 @@ export const ModeratorActions = ({classes, user, currentUser, refetch, setNotes,
   const signatureWithNote = (note:string) => {
     return `${signature}: ${note}\n`
   }
+  const handleNotes = () => {
+    if (notes != user.sunshineNotes) {
+      void updateUser({
+        selector: {_id: user._id},
+        data: {
+          sunshineNotes: notes
+        }
+      })
+    }
+  }
 
+  const signAndDate = (sunshineNotes:string) => {
+    if (!sunshineNotes.match(signature)) {
+      const padding = !sunshineNotes ? ": " : ": \n\n"
+      return signature + padding + sunshineNotes
+    }
+    return sunshineNotes
+  }
+  
+  const handleClick = () => {
+    const signedNotes = signAndDate(notes)
+    if (signedNotes != notes) {
+      setNotes(signedNotes)
+    }
+  }
   const handleReview = () => {
     if (canReview) {
       void updateUser({
@@ -100,6 +166,18 @@ export const ModeratorActions = ({classes, user, currentUser, refetch, setNotes,
     setNotes( newNotes )
   }
   
+  const handleNeedsReview = () => {
+    const newNotes = signatureWithNote("set to manual review") + notes;
+    void updateUser({
+      selector: {_id: user._id},
+      data: {
+        needsReview: true,
+        sunshineNotes: newNotes
+      }
+    })    
+    setNotes( newNotes )
+  }
+
   const banMonths = 3
   
   const handleBan = () => {
@@ -233,34 +311,81 @@ export const ModeratorActions = ({classes, user, currentUser, refetch, setNotes,
     }
 
     setNotes(newNotes);
-    // Refetch to ensure the button displays (toggled on/off) properly 
+    // If we have a efetch to ensure the button displays (toggled on/off) properly 
     refetch();
   }
 
-  return <div className={classes.row}>
-    <LWTooltip title="Snooze 10 (Appear in sidebar after 10 posts and/or comments)" placement="top">
-      <AddAlarmIcon className={classNames(classes.snooze10, classes.modButton)} onClick={() => handleSnooze(10)}/>
-    </LWTooltip>
-    <LWTooltip title="Snooze 1 (Appear in sidebar on next post or comment)" placement="top">
-      <SnoozeIcon className={classes.modButton} onClick={() => handleSnooze(1)}/>
-    </LWTooltip>
-    <LWTooltip title="Approve" placement="top">
-      <DoneIcon onClick={handleReview} className={classNames(classes.modButton, {[classes.canReview]: !classes.disabled })}/>
-    </LWTooltip>
-    <LWTooltip title="Ban for 3 months" placement="top">
-      <RemoveCircleOutlineIcon className={classes.modButton} onClick={handleBan} />
-    </LWTooltip>
-    <LWTooltip title="Purge (delete and ban)" placement="top">
-      <DeleteForeverIcon className={classes.modButton} onClick={handlePurge} />
-    </LWTooltip>
-    <LWTooltip title={user.sunshineFlagged ? "Unflag this user" : <div>
-      <div>Flag this user for more review</div>
-      <div><em>(This will not remove them from sidebar)</em></div>
-    </div>} placement="top">
-      <div onClick={handleFlag} className={classes.modButton} >
-        {user.sunshineFlagged ? <FlagIcon /> : <OutlinedFlagIcon />}
-      </div>
-    </LWTooltip>
+  return <div>
+    <div className={classes.notes}>
+      <Input
+        value={notes}
+        fullWidth
+        onChange={e => setNotes(e.target.value)}
+        onClick={e => handleClick()}
+        onBlur={handleNotes}
+        disableUnderline
+        placeholder="Notes for other moderators"
+        multiline
+        rowsMax={5}
+      />
+    </div>
+    <div className={classes.row}>
+      <LWTooltip title="Snooze 10 (Appear in sidebar after 10 posts and/or comments)" placement="top">
+        <AddAlarmIcon className={classNames(classes.snooze10, classes.modButton)} onClick={() => handleSnooze(10)}/>
+      </LWTooltip>
+      <LWTooltip title="Snooze 1 (Appear in sidebar on next post or comment)" placement="top">
+        <SnoozeIcon className={classes.modButton} onClick={() => handleSnooze(1)}/>
+      </LWTooltip>
+      <LWTooltip title="Approve" placement="top">
+        <DoneIcon onClick={handleReview} className={classNames(classes.modButton, {[classes.canReview]: !classes.disabled })}/>
+      </LWTooltip>
+      <LWTooltip title="Ban for 3 months" placement="top">
+        <RemoveCircleOutlineIcon className={classes.modButton} onClick={handleBan} />
+      </LWTooltip>
+      <LWTooltip title="Purge (delete and ban)" placement="top">
+        <DeleteForeverIcon className={classes.modButton} onClick={handlePurge} />
+      </LWTooltip>
+      <LWTooltip title={user.sunshineFlagged ? "Unflag this user" : <div>
+        <div>Flag this user for more in-depth review</div>
+        <div><em>(This will not remove them from sidebar)</em></div>
+      </div>} placement="top">
+        <div onClick={handleFlag} className={classes.modButton} >
+          {user.sunshineFlagged ? <FlagIcon /> : <OutlinedFlagIcon />}
+        </div>
+      </LWTooltip>
+      {!user.needsReview && <LWTooltip title="Return this user to the review queue">
+        <VisibilityOutlinedIcon className={classes.modButton} onClick={handleNeedsReview}/>
+      </LWTooltip>}
+    </div>
+    <div className={classes.row}>
+      <LWTooltip title={`${user.postingDisabled ? "Enable" : "Disable"} this user's ability to create posts`}>
+        <div className={classNames(classes.permissionsButton, {[classes.permissionDisabled]: user.postingDisabled})} onClick={handleDisablePosting}>
+          Posts
+        </div>
+      </LWTooltip>
+      <LWTooltip title={`${user.allCommentingDisabled ? "Enable" : "Disable"} this user's to comment (including their own shortform)`}>
+        <div className={classNames(classes.permissionsButton, {[classes.permissionDisabled]: user.allCommentingDisabled})} onClick={handleDisableAllCommenting}>
+          All Comments
+        </div>
+      </LWTooltip>
+      <LWTooltip title={`${user.commentingOnOtherUsersDisabled ? "Enable" : "Disable"} this user's ability to comment on other people's posts`}>
+        <div className={classNames(classes.permissionsButton, {[classes.permissionDisabled]: user.commentingOnOtherUsersDisabled})} onClick={handleDisableCommentingOnOtherUsers}>
+          Other Comments
+        </div>
+      </LWTooltip>
+      <LWTooltip title={`${user.conversationsDisabled ? "Enable" : "Disable"} this user's ability to start new private conversations`}>
+        <div className={classNames(classes.permissionsButton, {[classes.permissionDisabled]: user.conversationsDisabled})}onClick={handleDisableConversations}>
+          Conversations
+        </div>
+      </LWTooltip>
+    </div>
+    <div>
+      <LWTooltip title={`${mostRecentRateLimit?.active ? "Un-rate-limit" : "Rate-limit"} this user's ability to post and comment`}>
+        <div className={classes.permissionsButton}onClick={handleRateLimit}>
+          {MODERATOR_ACTION_TYPES[RATE_LIMIT_ONE_PER_DAY]}
+        </div>
+      </LWTooltip>
+    </div>
   </div>
 }
